Adreed/10.32.7 (#3542)
* 10.32.6 release / FIPS release (#3499)

* disable network iso

* Update offending packages

* Correct issues re: MSRC case #110341

* Alter intentional panics to return errors

* Changelog

* Add test to validate changes

* Satiate the clanker

* Error formatting

* version.go

* Update packages and add patch version

* update otel sdk

* Cover other open CVEs

* Remove 32-bit Windows ARM7 build

Broken in Golang, removed in next major revision anyway.

* Bump Go toolchain and security-relevant dependencies

- Go: 1.25.8 -> 1.25.11 (go.mod + CI/pipeline configs)

- golang.org/x/crypto: 0.47.0 -> 0.53.0

- golang.org/x/net: 0.49.0 -> 0.55.0

- github.com/go-jose/go-jose/v4: 4.1.3 -> 4.1.4 (indirect)

- google.golang.org/grpc: 1.79.1 -> 1.79.3 (indirect)

Transitive updates from go mod tidy: x/sync 0.19.0->0.21.0, x/sys 0.40.0->0.46.0, x/text 0.33.0->0.38.0.


* patch release

* stylistic changes from copilot :)


* go mod tidy

* Changelog

* Version

* TASK 38260338: Corrected a bug in the release pipeline. (#3500)

Apparently, in bash, Boolean pipeline parameters are converted into strings that begin with a capital letter - "True" and "False", not "true" and "false".

* TASK 38260338: Updated the release pipeline to sign the FIPS RPM files with the Mariner signing key. (#3501)

This is required to upload the files to the Azure Linux 3.0 repo at packages.microsoft.com. We'll need to make additional changes to the release pipeline to be able to upload those files to other Linux distros.

* Changelog comments

* sddl erroring changes

---------


* Generate resource.syso files during pipeline runs, embed into executable (#3479)

* Generate resource.syso files during pipeline runs, embed into executable

* Clanker satiation

* Actually the clanker was wrong

* Licenses

* Upgrade minio to better support more recent S3 regions (#3495)

* Upgrade minio to better support more recent S3 regions

* e2e test calls

* Correct other testing calls

* fully eliminate v6

* `go mod tidy`

* `go mod tidy` again

* Clanker satiation

* Clone map to increase safety with concurrency

* go mod -u (#3522)

* Fix SIGTERM cancellation during enumeration

SIGTERM cancellation could hang while enumeration was incomplete because the lifecycle manager synchronously prompted for stdin before scheduling job cancellation. Non-interactive callers such as timeout never supplied that response, so AzCopy remained running.

Replace the prompt with a non-resumability warning and continue cancellation immediately. Add focused unit coverage and a deterministic process-level E2E test that sends SIGTERM before enumeration starts and verifies AzCopy exits.

* Changelog, version

* comments

* Ensure flags that should be hidden are hidden

---------

+++ sddl/sddlHelper_linux.go
@@ -1127,7 +1127,7 @@ func getDaclString(sd []byte) (string, error) {
 
 	if !isRangeValid(dacloffset, 8, uint32(len(sd))) {
 		return "", fmt.Errorf("dacloffset (%d) points outside Security Descriptor of size %d bytes!",
-			dacloffset+8, len(sd))
+			uint64(dacloffset)+8, len(sd))
 	}
 
 	// ACL.AclRevision.
